Got unity project working again
