+++ UnityGLTF/Assets/GLTF/Examples/GLTFExporterTest.cs
@@ -1,21 +1,16 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using GLTF;
+using UnityGLTFSerialization;
 using System.IO;
 
 public class GLTFExporterTest : MonoBehaviour {
 
-	// Use this for initialization
-	void Awake () {
-		var exporter = new GLTFExporter(new []{ transform });
-		var appPath = Application.dataPath;
-		var wwwPath = appPath.Substring(0, appPath.LastIndexOf("Assets")) + "www";
-		exporter.SaveGLTFandBin(Path.Combine(wwwPath, "TestScene"), "TestScene");
-	}
-
-	// Update is called once per frame
-	void Update () {
-
-	}
+    // Use this for initialization
+    void Awake () {
+        var exporter = new GLTFSceneExporter(new []{ transform });
+        var appPath = Application.dataPath;
+        var wwwPath = appPath.Substring(0, appPath.LastIndexOf("Assets")) + "www";
+        exporter.SaveGLTFandBin(Path.Combine(wwwPath, "TestScene"), "TestScene");
+    }
 }
