+++ UnityGLTF/Assets/GLTF/Examples/WebServerComponent.cs
@@ -2,17 +2,17 @@
 
 public class WebServerComponent : MonoBehaviour {
 
-	private SimpleHTTPServer _server;
+    private SimpleHTTPServer _server;
 
-	void Start () {
-		var appPath = Application.dataPath;
-		var wwwPath = appPath.Substring(0, appPath.LastIndexOf("Assets")) + "www";
-		_server = new SimpleHTTPServer(wwwPath, 8080);
-		Debug.Log("Starting web server...");
-	}
+    void Start () {
+        var appPath = Application.dataPath;
+        var wwwPath = appPath.Substring(0, appPath.LastIndexOf("Assets")) + "www";
+        _server = new SimpleHTTPServer(wwwPath, 8080);
+        Debug.Log("Starting web server...");
+    }
 
-	private void OnDestroy()
-	{
-		_server.Stop();
-	}
+    private void OnDestroy()
+    {
+        _server.Stop();
+    }
 }
