+++ UnityGLTF/Assets/GLTF/Scripts/CacheData/MaterialCacheData.cs
@@ -1,0 +1,14 @@
+﻿namespace UnityGLTFSerialization.CacheData
+{
+    public class MaterialCacheData
+    {
+        public UnityEngine.Material UnityMaterial { get; set; }
+        public UnityEngine.Material UnityMaterialWithVertexColor { get; set; }
+        public GLTFSerialization.Material GLTFMaterial { get; set; }
+
+        public UnityEngine.Material GetContents(bool useVertexColors)
+        {
+            return useVertexColors ? UnityMaterialWithVertexColor : UnityMaterial;
+        }
+    }
+}
