--- UnityGLTF/Assets/GLTF/Scripts/GLTFExporter.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: acf52387dda9a6a449902d25a54d25c0
-timeCreated: 1494288226
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
