--- UnityGLTF/Assets/GLTF/Scripts/GLTFLoader.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 714ac887a6a96814db73f4f557dc5bfb
-timeCreated: 1494369866
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
