+++ UnityGLTF/Assets/GLTF/Scripts/Tests/Editor/GLTFBenchmark.cs
@@ -2,7 +2,7 @@
 using UnityEngine;
 using UnityEditor;
 using System.Collections;
-using GLTF;
+using GLTFSerialization;
 using UnityEngine.Networking;
 using System.Text;
 using System.IO;
@@ -38,7 +38,7 @@ public class GLTFBenchmark : MonoBehaviour
 			for (var i = 0; i < NumberOfIterations; i++)
 			{
 				timer.Start();
-				GLTFParser.ParseString(www.downloadHandler.text);
+				GLTFParser.ParseJson(Encoding.ASCII.GetBytes(www.downloadHandler.text));
 				timer.Stop();
 
 				Debug.LogFormat("Iteration {0} took: {1}ms", i, timer.ElapsedMilliseconds);
