+++ UnityGLTF/Assets/GLTF/Scripts/Tests/Integration/GLTFTestComponent.cs
@@ -1,23 +1,23 @@
 ﻿using UnityEngine;
 using System.Collections;
-using GLTF;
+using UnityGLTFSerialization;
 
 public class GLTFTestComponent : MonoBehaviour {
-	public string Url;
-	public bool Multithreaded = true;
+    public string Url;
+    public bool Multithreaded = true;
 
-	public Shader GLTFStandard;
+    public Shader GLTFStandard;
 
 
-	IEnumerator Start()
-	{
-		var loader = new GLTFLoader(
-				Url,
-				gameObject.transform
-			);
-		loader.SetShaderForMaterialType(GLTFLoader.MaterialType.PbrMetallicRoughness, GLTFStandard);
-		loader.Multithreaded = Multithreaded;
-		yield return loader.Load();
-		IntegrationTest.Pass();
-	}
+    IEnumerator Start()
+    {
+        var loader = new GLTFSceneImporter(
+            Url,
+            gameObject.transform
+            );
+
+        loader.SetShaderForMaterialType(GLTFSceneImporter.MaterialType.PbrMetallicRoughness, GLTFStandard);
+        yield return loader.Load(-1, Multithreaded);
+        IntegrationTest.Pass();
+    }
 }
